+++ 3_ProcessModel/src/processmodel/simplest/WorkshopProduceMethod.java
@@ -50,7 +50,7 @@ public class WorkshopProduceMethod extends SimpleMethod {
     }
 
     @Override
-    public int getWeight() {
+    public float getWeight() {
         return getWeightA(coeffSimp);
     }
